+++ Scripts/Items/Artifacts/Decorative/ShipsBellOfBmvArarat.cs
