+++ Scripts/Items/Artifacts/Decorative/SternAnchorOfBmvArarat.cs
