+++ Scripts/Items/Artifacts/Equipment/Weapons/Abhorrence.cs
