+++ Scripts/Scripts.csproj
@@ -652,17 +652,14 @@
     <Compile Include="Items\Artifacts\Equipment\Weapons\TheDeceiver.cs" />
     <Compile Include="Items\Decorative\TatteredWallMap.cs" />
     <Compile Include="Items\Equipment\Armor\BoneArms.cs" />
-    <Compile Include="Items\Equipment\Weapons\Craven.cs" />
-    <Compile Include="Items\Equipment\Weapons\GargishCaptainJohnesBlade.cs" />
-    <Compile Include="Items\Equipment\Weapons\CaptainJohnesBlade.cs" />
-    <Compile Include="Items\Equipment\Weapons\Abhorrence.cs" />
-    <Compile Include="Items\Equipment\Weapons\GargishPincer.cs" />
-    <Compile Include="Items\Decorative\FigureheadOfBmvArarat.cs" />
-    <Compile Include="Items\Decorative\ShipsBellOfBmvArarat.cs" />
-    <Compile Include="Items\Decorative\SternAnchorOfBmvArarat.cs" />
-    <Compile Include="Items\Equipment\Weapons\Pincer.cs" />
-    <Compile Include="Items\Equipment\Weapons\GargishEquivocation.cs" />
-    <Compile Include="Items\Equipment\Weapons\Equivocation.cs" />
+    <Compile Include="Items\Artifacts\Equipment\Weapons\Craven.cs" />
+    <Compile Include="Items\Artifacts\Equipment\Weapons\CaptainJohnesBlade.cs" />
+    <Compile Include="Items\Artifacts\Equipment\Weapons\Abhorrence.cs" />
+    <Compile Include="Items\Artifacts\Decorative\FigureheadOfBmvArarat.cs" />
+    <Compile Include="Items\Artifacts\Decorative\ShipsBellOfBmvArarat.cs" />
+    <Compile Include="Items\Artifacts\Decorative\SternAnchorOfBmvArarat.cs" />
+    <Compile Include="Items\Artifacts\Equipment\Weapons\Pincer.cs" />
+    <Compile Include="Items\Artifacts\Equipment\Weapons\Equivocation.cs" />
     <Compile Include="Items\Artifacts\Equipment\Weapons\FNPitchfork.cs" />
     <Compile Include="Items\Books\RunicAtlas.cs" />
     <Compile Include="Items\Consumables\AlterContract.cs" />
